refactor(web): extract dockview measure helper, fix observer leak + stale serialize
Address review feedback (CodeRabbit + claude bot):

- Extract `measureDockviewContainer` to `lib/state/dockview-measure.ts`
  so `dockview-store.ts` and `dockview-layout-restore.ts` share one
  implementation.
- Replace the `__safe`-named two-line capture pattern with the cleaner
  `const { width: safeWidth, height: safeHeight } = …` destructure used
  elsewhere in the file.
- Track the `setupContainerResizeSync` and `setupGroupTracking` disposers
  in a `readyDisposersRef` and call them in the unmount cleanup so the
  ResizeObserver and event listeners disconnect on navigate-away rather
  than leaking until tab close.
- `saveOutgoingEnv`: when serializing `preMaximizeLayout` to the env
  slot, use the measured container size instead of `api.width/height`,
  for the same reason as the other call sites — a drifted internal width
  shouldn't get persisted into the saved layout.
